Add input and textarea UI components + contact page improvements (#114)
* feat(ui): Add default input and textarea

* feat(ui): Customize input and textarea

Applies SESA styling to the shadcn/ui components.

* refactor(contact): Use shadcn/ui for contact page

Changes the "Contact Us" page to use the newly-added shadcn/ui input and
textare components.

* feat(contact): Add `autocomplete` to inputs

A general UX improvement. Not exactly related to the PR, but thought I'd
fix it while here. This allows password managers to more easily
determine what each input is for and autocomplete it.