Rename all pages to lowercase (#161)
* refactor: Rename all pages to lowercase

Primarily an aesthetic change. Renames all pages that had
PascalCase-type names to be snake_case, ideally just a single word.

Links to the old pages have also been replaced.

The old `resources/[id]` dynamic page was moved to `resource/[id]`. It
would make more sense to move this under the main `resources` page, but
that can be done in a separate PR.

* refactor: Rename `resource/[id]` -> `resources/[id]`

This was a much simpler change than I thought it might be, so I'll just
add it in this PR.

* Updated link for faq in en.json

---------

Co-authored-by: tazim04 <35643104+tazim04@users.noreply.github.com>