Use infinite scroll for resources (#256)
* refactor: Rename `getPage` -> `getOffsetPage`

Will help disambiguate between cursor-based and offset-based pagination.

* feat(resources): Create procedure for cursor-based pagination

* feat(resources): Use infinite scroll

* feat(resources): Update prefetches

* feat(resources): Make page size constant

Can help prevent issues related to accidentally mixing and matching page
sizes.

* fix(resources): Add missing pagination functions

* feat(resources): Use consistent default parameters

Allows the query cache to be properly reused between pages.

* chore: Remove debug log

* fix(resources): Update order for offset pagination

* feat(resources): Make prefetch more eager on mobile

Makes it less likely that the user will hit the end of the list before
the next page is ready.

* feat(resources): Remove dropdown for number of rows

* fix(resources): Tweak cursor-related code

* feat(resources): Make desktop prefetch less aggressive

* fix(resources): Fix cursor pagination

A slight off-by-one error. Essentially, the cursor comparison uses a
strict inequality, but the first element of the next page was being used
as the cursor, meaning that the first element of every next page ended
up accidentally being skipped.

* feat(resources): Change from auto-load to "Load More"

* feat(resources): Add message for end of list

* chore: Apply Biome fixes

* fix(resources): Fix "Load More" on mobile